refactor(backend): move stubbed-app fixtures into conftest
The app_with_stubbed_agent and client fixtures were defined inline in
test_assistant_endpoints.py, which meant any other test module wanting a
stubbed assistant app had to import fixtures across modules or rebuild the
whole thing. Moved them to tests/conftest.py so pytest hands them out
normally. Pure move -- no behavior change, SECRET went along with them.
